Move `_get_available_gpu_memory()` to `cupy_utils.py`
