+++ httomo/task_runner.py
@@ -33,6 +33,7 @@ from httomo.utils import (
     log_once,
     remove_ansi_escape_sequences,
 )
+from httomo.cupy_utils import _get_available_gpu_memory
 from httomo.wrappers_class import HttomolibWrapper, HttomolibgpuWrapper, TomoPyWrapper
 from httomo.yaml_utils import open_yaml_config
 
@@ -736,22 +737,6 @@ def _determine_platform_sections(
     )
 
     return ret
-
-
-def _get_available_gpu_memory(safety_margin_percent: float = 10.0) -> int:
-    try:
-        import cupy as cp
-
-        dev = cp.cuda.Device()
-        # first, let's make some space
-        pool = cp.get_default_memory_pool()
-        pool.free_all_blocks()
-        cache = cp.fft.config.get_plan_cache()
-        cache.clear()
-        available_memory = dev.mem_info[0] + pool.free_bytes()
-        return int(available_memory * (1 - safety_margin_percent / 100.0))
-    except:
-        return int(100e9)  # arbitrarily high number - only used if GPU isn't available
 
 
 def _update_max_slices(
